CSV Export files break when headers contain new line characters
If a column header contains a new line (which we use to keep column widths narrow), then the exported CSV files contain new lines in their headings which breaks the CSV format.   This fix replaces all whitespace characters and multiple whitespace strings with single spaces to fix that.

- Fix vscode merging corruption non-sense

Fixing merge